Add cleanup option New GUI Initial code to check for busybox status

+++ trunk/src/me/timos/busyboxonrails/ServiceNormalOperation.java
@@ -6,13 +6,12 @@ import static me.timos.busyboxonrails.Utility.shellExec;
 import java.io.File;
 
 import me.timos.br.Logcat;
-import me.timos.busyboxonrails.R;
 import android.widget.Toast;
 
-public class ServiceNormalInstall extends ServiceBase {
+public class ServiceNormalOperation extends ServiceBase {
 
-	public ServiceNormalInstall() {
-		super(ServiceNormalInstall.class.getName());
+	public ServiceNormalOperation() {
+		super(ServiceNormalOperation.class.getName());
 	}
 
 	@Override
@@ -52,28 +51,34 @@ public class ServiceNormalInstall extends ServiceBase {
 				"if [ \"`./busybox ls -l $i|./busybox grep busybox`\" ]; then",
 				"echo $i", "rm $i", "fi", "done", "rm /system/xbin/busybox");
 
-		Logcat.d("Write busybox to /system/xbin");
-		shellExec(tmp.getParent(), null, "cat " + tmp + " > " + target,
-				"chmod 755 " + target);
-		if (!target.canExecute()
-				|| !checkFileIntegrity(this, target, getBusyboxResPath())) {
-			Logcat.e("ERROR WRITE BUSYBOX TO /system/xbin");
-			mApp.showToast(R.string.error_write_busybox_to_xbin,
-					Toast.LENGTH_LONG);
-			return;
-		}
+		if (mOpId == R.id.radCleanupInstall) {
+			Logcat.d("Write busybox to /system/xbin");
+			shellExec(tmp.getParent(), null, "cat " + tmp + " > " + target,
+					"chmod 755 " + target);
+			if (!target.canExecute()
+					|| !checkFileIntegrity(this, target, getBusyboxResPath())) {
+				Logcat.e("ERROR WRITE BUSYBOX TO /system/xbin");
+				mApp.showToast(R.string.error_write_busybox_to_xbin,
+						Toast.LENGTH_LONG);
+				return;
+			}
 
-		Logcat.d("Create applets");
-		String ret = shellExec(target.getParent(), null,
-				"for i in `./busybox --list`; do", "rm $i",
-				"./busybox ln -s busybox $i", "done", "./busybox ls -l "
-						+ lastApplet);
-		if (ret.endsWith("zcat -> busybox")) {
-			Logcat.d("SUCCEEDED");
-			mApp.showToast(R.string.msg_install_succeeded, Toast.LENGTH_LONG);
+			Logcat.d("Create applets");
+			String ret = shellExec(target.getParent(), null,
+					"for i in `./busybox --list`; do", "rm $i",
+					"./busybox ln -s busybox $i", "done", "./busybox ls -l "
+							+ lastApplet);
+			if (ret.endsWith("zcat -> busybox")) {
+				Logcat.d("INSTALLATION SUCCEEDED");
+				mApp.showToast(R.string.msg_install_succeeded,
+						Toast.LENGTH_LONG);
+			} else {
+				mApp.showToast(R.string.error_applets, Toast.LENGTH_LONG);
+				Logcat.e("FAILED creating applets \n" + ret);
+			}
 		} else {
-			mApp.showToast(R.string.error_applets, Toast.LENGTH_LONG);
-			Logcat.e("FAILED creating applets \n" + ret);
+			Logcat.d("UNINSTALL COMPLETED");
+			mApp.showToast(R.string.msg_uninstall_completed, Toast.LENGTH_LONG);
 		}
 
 		Logcat.d("Remount /system readonly");
